requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
awesome-slugify==1.6.5
Babel==2.6.0
backcall==0.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
decorator==4.3.2
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.35.0
django-appconf==1.0.3
django-bootstrap3==11.0.0
django-bootstrap4==0.0.6
django-braces==1.13.0
django-crispy-forms==1.7.2
django-datetime-widget==0.9.3
django-environ==0.4.4
django-model-utils==3.1.1
django-select2==6.0.2
django-simple-history==2.3.0
django-tenants==2.0.0
djangorestframework==3.8.2
docutils==0.14
idna==2.8
imagesize==1.1.0
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
MarkupSafe==1.1.1
oauthlib==3.0.1
packaging==19.0
parso==0.3.4
pickleshare==0.7.5
Pillow==5.1.0
prompt-toolkit==2.0.9
psycopg2==2.7.7
pycosat==0.6.3
pycparser==2.19
Pygments==2.3.1
pyOpenSSL==18.0.0
pyparsing==2.3.1
PySocks==1.6.8
python3-openid==3.1.0
pytz==2018.4